Create DPVWriter spec.
The idea is to move all the tests that deal without output formatting into the writer spec.
